% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bartext.R
\name{geom_bartext}
\alias{geom_bartext}
\title{Labeled Bar Charts}
\usage{
geom_bartext(mapping = NULL, data = NULL, position = "stack", ...,
  width = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, parse = FALSE)
}
\description{
This function adds a geom_bartext geom which essentially composes geom_bar() and geom_text()
}
\examples{

library(dplyr)
library(ggplot2)

mpg_manu <- mpg \%>\%
    distinct(manufacturer, model)

ggplot(mpg_manu, aes(x = manufacturer)) +
    geom_bartext(aes(label = model))

ggplot(mpg_manu, aes(x = manufacturer)) +
    geom_bartext(aes(label = model), colour = "red", textcolour = "black", fill = "white")
}
